Feat/new UI edit observation attachment (#372)
* redesign observation edit screen

* update icons, create separate components for ActionTab

* Update ActionTab component

* change translations in the audio button

* update description field edit observation screen

* fixed problem with a feature flag

* update thumbnail components

* update imports and rename the folder with icons

* change structure files, add play icon

* change the type of AudioThumbnail

---------
